expose an api
I need this to accomodate unhandledRejections in AVA.

I do not intend to document it immediately.
